fix1

+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/ApplicationRepository.java
@@ -3,5 +3,5 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationRepository extends JpaRepository<Application, Long> {
+public interface ApplicationRepository extends JpaRepository<Application, Integer> {
 }
