+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/CustomerRepository.java
@@ -3,5 +3,7 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+import java.util.List;
+
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
