+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/GoodsRepository.java
@@ -3,5 +3,5 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.Goods;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoodsRepository  extends JpaRepository<Goods, Long> {
+public interface GoodsRepository  extends JpaRepository<Goods, Integer> {
 }
