+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/OrderExecutorsRepository.java
@@ -3,5 +3,5 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.OrderExecutor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderExecutorsRepository extends JpaRepository<OrderExecutor, Long> {
+public interface OrderExecutorsRepository extends JpaRepository<OrderExecutor, Integer> {
 }
