+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/RouteRepository.java
@@ -3,5 +3,7 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RouteRepository extends JpaRepository<Route, Long> {
+import java.util.List;
+
+public interface RouteRepository extends JpaRepository<Route, Integer> {
 }
