+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/TransportRepository.java
@@ -3,5 +3,7 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.Transport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransportRepository extends JpaRepository<Transport, Long> {
+import java.util.List;
+
+public interface TransportRepository extends JpaRepository<Transport, Integer> {
 }
