+++ salesDepartmentSystem/src/main/java/egorov/com/salesdepartmentsystem/repositories/TypeOfGoodsRepository.java
@@ -3,5 +3,7 @@ package egorov.com.salesdepartmentsystem.repositories;
 import egorov.com.salesdepartmentsystem.models.TypeOfGoods;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeOfGoodsRepository extends JpaRepository<TypeOfGoods, Long> {
+import java.util.List;
+
+public interface TypeOfGoodsRepository extends JpaRepository<TypeOfGoods, Integer> {
 }
